+++ lib/generators/rpush_migration_generator.rb
@@ -51,6 +51,7 @@ class RpushMigrationGenerator < Rails::Generators::Base
     add_rpush_migration('rpush_3_3_0_updates')
     add_rpush_migration('rpush_3_3_1_updates')
     add_rpush_migration('rpush_4_1_0_updates')
+    add_rpush_migration('rpush_4_1_1_updates')
   end
 
   protected