+++ lib/rpush/apns_feedback.rb
@@ -7,6 +7,7 @@ module Rpush
       # Redis stores every App type on the same namespace, hence the
       # additional filtering
       next unless app.service_name == 'apns'
+      next unless app.feedback_enabled
 
       receiver = Rpush::Daemon::Apns::FeedbackReceiver.new(app)
       receiver.check_for_feedback