+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
-## Unreleased
+## 4.1.1 (2019-05-13)
 
-- Allow disabling of APNS feedback for specific rpush apps [#491](https://github.com/rpush/rpush/pull/491) (by [@drn](https://github.com/drn)).
+### Added
+
+- Allow disabling of APNS feedback for specific Rpush apps [#491](https://github.com/rpush/rpush/pull/491) (by [@drn](https://github.com/drn)).
+
+### Changed
+
 - Switch from ANSI to Rainbow. ([#496](https://github.com/rpush/rpush/pull/496) by [@drn](https://github.com/drn))
 
 ## 4.1.0 (2019-04-17)